refactored interface display in terminal

+++ hangman.rb
@@ -16,26 +16,27 @@ class Noose
 
   def turns_left
     @turn -= 1
-    puts "\n\n"
-    puts "#{@turn} guesses remaining!\n\n"
+    puts "#{@turn} guesses remaining!\n"
     @turn
   end
 
   def guessed(guess)
+    system "clear"
+    puts "Your guess: #{guess}"
     if guess.length > 1
       if guess == @word
         puts
-        puts "You've won!"
+        puts "You've won!\n\n"
         @turn = 1
       else
-        puts "\nNope! Try again!"
+        puts "\nNope! Try again!\n\n"
       end
     else
       if /#{guess}/.match(@word) != nil
-        puts "\nA match!"
+        puts "\nA match!\n\n"
         update_word(guess)
       else
-        puts "\nNope! Try again!"
+        puts "\nNope! Try again!\n\n"
         update_wrong(guess)
       end
     end
@@ -61,7 +62,7 @@ class Noose
     end
     if @word == @word_progress
       puts
-      puts "You've won!"
+      puts "You've won!\n"
       @turn = 1
     end
   end
@@ -79,23 +80,25 @@ def save_game(noose_object)
 end
 
 def load_game
-  #if exist?("data/save.yaml")
+  if File.exist?("data/save.yaml")
     save_file = File.open("data/save.yaml")
     yaml = save_file.read
-    noose = YAML::load(yaml)
-    noose
-  #else
-    #puts "No saved file detected. Starting new game."
-    #chosen_word = get_word
-    #noose = Noose.new(chosen_word, chosen_word.length + 2)
-  #end
+    noose = YAML::load(yaml)  
+  else
+    puts "No saved file detected. Starting new game.\n"
+    chosen_word = get_word
+    noose = Noose.new(chosen_word, chosen_word.length + 2)
+  end
+  noose
 end
 
 def display_main_menu
+  system "clear"
   puts "1. Play!"
   puts "2. Load Save"
   print "Enter # option: "
   option = gets.chomp
+  system "clear"
   return option
 end
 
@@ -103,7 +106,7 @@ def guess_input(guess, noose)
   guess = guess.to_s.downcase
   if guess == "save"
     save_game(noose)
-    puts "Game saved!"
+    puts "\nGame saved!\n"
     puts "Enter your guess(single letter or entire word):" 
     guess = guess_input(gets.chomp, noose)
   end
@@ -137,7 +140,7 @@ while !exit
 
   if noose.turns_left > 0
     puts
-    puts "Enter 'save' into prompt to save.\n"
+    puts "Enter 'save' into prompt to save.\n\n"
     print "Hangman: "
     noose.display_word_progress
     puts
@@ -150,7 +153,7 @@ while !exit
   else
     print "Answer: "
     noose.display_answer 
-    puts "Game over!"
+    puts "\nGame over!\n"
     exit = true
   end
 end
